refactor: move patient age calculation and date formatting to utils
