+++ apps/ehr/src/components/PatientDateOfBirth.tsx
@@ -1,7 +1,8 @@
 import { Typography } from '@mui/material';
 import { Box, useTheme } from '@mui/system';
 import { ReactElement } from 'react';
-import { calculatePatientAge, formatDateUsingSlashes } from '../helpers/formatDateTime';
+import { formatDateUsingSlashes } from '../helpers/formatDateTime';
+import { calculatePatientAge } from 'utils';
 
 export const PatientDateOfBirth = ({ dateOfBirth }: { dateOfBirth: string }): ReactElement => {
   const theme = useTheme();
