+++ apps/ehr/src/features/css-module/parser/constants.ts
@@ -1,7 +1,5 @@
 export const WEIGHT_EXTENSION_URL = 'https://fhir.zapehr.com/r4/StructureDefinitions/weight';
 export const WEIGHT_LAST_UPDATED_EXTENSION_URL = 'https://fhir.zapehr.com/r4/StructureDefinitions/weight-last-updated';
-export const DATE_FORMAT = 'yyyy-MM-dd';
-export const DISPLAY_DATE_FORMAT = 'MM/dd/yyyy';
 export const LBS_TO_KG_FACTOR = 0.45359237;
 export const WEIGHT_ROUNDING_PRECISION = 100;
