+++ packages/utils/lib/utils/dateUtils.ts
@@ -293,3 +293,51 @@ export const distributeTimeSlots = (
 
   return availableSlots;
 };
+
+export const DATE_FORMAT = 'yyyy-MM-dd';
+export const DISPLAY_DATE_FORMAT = 'MM/dd/yyyy';
+
+export function calculatePatientAge(date: string | null | undefined): string | null | undefined {
+  if (!date) {
+    return date;
+  }
+
+  const INVALID_DATE_FALLBACK = '';
+  const dob = DateTime.fromISO(date);
+
+  if (!dob.isValid) {
+    return INVALID_DATE_FALLBACK;
+  }
+
+  const now = DateTime.now();
+
+  if (dob > now) {
+    return '0 d';
+  }
+
+  const MINIMUM_AGE_IN_YEARS_TO_DISPLAY_YEARS = 2;
+  const MINIMUM_AGE_IN_MONTHS_TO_DISPLAY_MONTHS = 2;
+
+  const diff = now.diff(dob, ['years', 'months', 'days']);
+  const { years, months } = diff.toObject();
+
+  if (years && years >= MINIMUM_AGE_IN_YEARS_TO_DISPLAY_YEARS) {
+    return `${Math.floor(years)} y`;
+  }
+
+  const fullMonths = (years || 0) * 12 + (months || 0);
+
+  if (fullMonths >= MINIMUM_AGE_IN_MONTHS_TO_DISPLAY_MONTHS) {
+    return `${Math.floor(fullMonths)} m`;
+  }
+
+  const fullDays = now.diff(dob).as('days');
+  return `${Math.floor(fullDays)} d`;
+}
+
+export const formatDOB = (birthDate: string | undefined): string | undefined => {
+  if (!birthDate) return undefined;
+  const birthday = DateTime.fromFormat(birthDate, DATE_FORMAT).toFormat(DISPLAY_DATE_FORMAT);
+  const age = calculatePatientAge(birthDate);
+  return `${birthday} (${age})`;
+};
